requirements.txt
openai
groq
sentence-transformers
faiss-cpu
python-dotenv
langchain
langchain-groq 